requirements.txt
fastapi
pyodbc
python-decouple
uvicorn
